Initial commit.
Readme briefly explains the purpose of IronBank as well as lists some of the services it will support.

main.rs contains the setup of the application, this should be a place for setting up the webserver and the database etc, but no actual business logic.

context_filters and response_filters are for plumbing of requests.

Added a basic context filter which adds an Accepts:application/json header to requests with no Accepts header set (just a sensible default for an API)

Begun work on a response filter which encodes the response into JSON (if the Accepts:application/json header is set). We'll need more of these filters, one for each data type which we support. Not yet decided how the filter gets hold of a PORO (Plain Old Rust Object) which it can serialize as JSON)

Signed-off-by: Martin Evans <martindevans@gmail.com>